refactor: split Editor component into Editor and EditorWrapper
