+++ templates/frontend/src/components/shared/editor-wrapper/index.js
@@ -1,0 +1,1 @@
+export { default } from './editor-wrapper.jsx';
