+++ templates/frontend/src/components/shared/template-form/view-only-form.jsx
@@ -14,7 +14,7 @@ import Files from './files';
 import PromptsReadOnly from './prompts/prompts-read-only';
 import styles from './template-form.module.scss';
 
-const Editor = dynamic(import('components/shared/editor'), { ssr: false });
+const Editor = dynamic(import('components/shared/editor-wrapper'), { ssr: false });
 
 const cx = classNames.bind(styles);
